Refactor CacheClearAllCommand testing
Replaced integration test with a unit test to improve test isolation and efficiency. Removed verbose output from CacheClearAllCommand for simplification.
